Added user entity, fixes #113
This commit adds a user enity and DAO. The user has a username and moodcat Id.
I have also added fields for soundcloud id and soundcloud oauth tokens.
This way we can persist the soundcloud login with a cookie, rather than require a new login on each page refresh.
I've also bounded the author field for chat messages to the user.
We might want to debate on wether or not this is desirable.
Pros: no possible username manipulation for the chat, the username is enforced by the relational model.
Cons: we will have to refactor the frontend and API call in order to merge in this PR (basically you should have a user attached to the session).

For this PR to work in production the database model needs to be updated accordingly.
The test database (h2) however works (and the bootstrapper is updated as well)